requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
scikit-learn==1.5.1
numpy==2.1.0
pandas==2.2.2
streamlit==1.38.0
paho-mqtt==2.1.0
matplotlib==3.9.2
joblib==1.4.2
